chore: fixing clippy issues

+++ crates/argon-codegen-wasm/src/lib.rs
@@ -1124,7 +1124,7 @@ impl<'a> ModuleLowerer<'a> {
                     wasm_fn
                         .instruction(&Instruction::LocalGet(value_local(ctx.params_count, *arg)));
                     wasm_fn
-                        .instruction(&Instruction::I32Store(memarg((4 + idx as u64 * 4) as u64)));
+                        .instruction(&Instruction::I32Store(memarg(4 + idx as u64 * 4)));
                 }
 
                 ctx.set_value_shape(*dest, Some(HeapValueShape::Object(layout)));
